Committing again because of mistake made with last commit. As stated, renamed all occurrences of TelemData to Pixhawk and all occurrences of telemData to pixhawk.
